+++ lib/src/properties/base_popup_props.dart
@@ -1,11 +1,12 @@
 import 'package:dropdown_search/src/base_dropdown_search.dart';
-import 'package:dropdown_search/src/properties/dropdown_props.dart';
 import 'package:dropdown_search/src/properties/infinite_scroll_props.dart';
 import 'package:dropdown_search/src/properties/list_view_props.dart';
 import 'package:dropdown_search/src/properties/scrollbar_props.dart';
 import 'package:dropdown_search/src/properties/suggestions_props.dart';
 import 'package:dropdown_search/src/properties/text_field_props.dart';
 import 'package:flutter/material.dart';
+
+import 'click_props.dart';
 
 abstract class BasePopupProps<T> {
   /// popup title